model/requirements.txt
PyYAML
Pillow
numpy
torch
matplotlib
scipy